refactor(proxy): move out transaction creation logic into a separate module
