+++ proxy/Cargo.toml
@@ -42,9 +42,9 @@ alloy-consensus = { git = "https://github.com/alloy-rs/alloy", rev = "dd7a999", 
 alloy-rlp = "0.3.4"
 
 [dev-dependencies]
-alloy-signer = {git = "https://github.com/alloy-rs/alloy", rev = "dd7a999"}# "0.1.3"
-alloy-signer-wallet= {git = "https://github.com/alloy-rs/alloy", rev = "dd7a999"}# "0.1.3"
-alloy-network = {git = "https://github.com/alloy-rs/alloy", rev = "dd7a999"}# "0.1.3"
+alloy-signer = { git = "https://github.com/alloy-rs/alloy", rev = "dd7a999" }
+alloy-signer-wallet= { git = "https://github.com/alloy-rs/alloy", rev = "dd7a999" }
+alloy-network = { git = "https://github.com/alloy-rs/alloy", rev = "dd7a999" }
 base64 = "0.21"
 bs58 = "0.5.1"
 solana-program-test = "1.17.34"
